Add mobile side navigation menu for `users`
In this commit, I have introduced a responsive side navigation menu,
specifically designed for mobile devices. Prior to this change,
our application lacked a side navigation menu for smaller screens,
making it less user-friendly for mobile users. This addition enhances
the mobile user experience by providing a convenient and intuitive way
to navigate through our application.

Please note that this is the initial implementation, and we will
continue to refine it in the future to make it even better.

The new side navigation menu can be accessed by clicking on a menu icon.

------------------------------------------------------------------------

	new file:   user/nav.js
	modified:   user/nav.php
	modified:   user/style/_nav.scss